requirements.txt
fastapi==0.113.0
uvicorn[standard]==0.25.0
sqlalchemy==2.1.0
pydantic==2.8.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.2
aiomysql==0.1.1
pytest==8.4.3
httpx==0.28.3
